create audio unit tests
