+++ docs/TEST_CASES.md
@@ -1,0 +1,7 @@
+# Test Cases
+
+## Audio Testing
+
++ Sine equation returns 0
++ Sine equation returns expected value
++ Generate test samples
