+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy==2.2.2
+scipy==1.10.1
+sounddevice==0.5.1
